Port shielding middleware to TypeScript (#51146)
