--- src/shielding/middleware/index.js
@@ -1,21 +1,0 @@
-import express from 'express'
-
-import handleInvalidQuerystrings from './handle-invalid-query-strings.js'
-import handleInvalidPaths from './handle-invalid-paths.js'
-import handleOldNextDataPaths from './handle-old-next-data-paths.js'
-import handleInvalidQuerystringValues from './handle-invalid-query-string-values.js'
-import handleInvalidNextPaths from './handle-invalid-nextjs-paths.js'
-import handleInvalidHeaders from './handle-invalid-headers.js'
-import rateLimit from './rate-limit.js'
-
-const router = express.Router()
-
-router.use(rateLimit)
-router.use(handleInvalidQuerystrings)
-router.use(handleInvalidPaths)
-router.use(handleOldNextDataPaths)
-router.use(handleInvalidQuerystringValues)
-router.use(handleInvalidNextPaths)
-router.use(handleInvalidHeaders)
-
-export default router
