+++ src/shielding/tests/invalid-querystring-values.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest'
 
-import { get } from '#src/tests/helpers/e2etest.js'
+import { get } from '@/tests/helpers/e2etest.js'
 
 describe('invalid query string values', () => {
   test.each(['platform', 'tool'])('%a key', async (key) => {
